refactor(profile): rename Services.Profile → Services.Profiles (match plural folder convention)
The Services folder convention is plural where the section represents a
collection of entities (Camps, Teams, Tickets, Notifications, Shifts, …).
Services.Profile was the outlier — singular, while the matching interfaces
namespace (Humans.Application.Interfaces.Profiles) was already plural.

Renames Services/Profile/ → Services/Profiles/ for the 7 services
(AccountMergeService, CommunicationPreferenceService, ContactFieldService,
DuplicateAccountService, EmailProblemsService, ProfileService, UserEmailService)
and rewrites all callers across src/ and tests/. Updates the
NoObsoleteNavReads baseline path entries.

Now that the drift is fixed:
- Drops the temporary [Grandfathered("HUM0012", ...)] attributes (and the
  `using Humans.Application.Architecture;` directives that were added solely
  for those attributes) from the 7 services.
- Removes HUM0012 from Directory.Build.props WarningsNotAsErrors — the rule
  is now error-only across the codebase.

Tests: 2930 pass / 1 skip.
