+++ src/Humans.Application/Services/Profiles/AccountMergeService.cs
@@ -1,7 +1,6 @@
 using System.Transactions;
 using Microsoft.Extensions.Logging;
 using NodaTime;
-using Humans.Application.Architecture;
 using Humans.Application.Extensions;
 using Humans.Application.Interfaces.Gdpr;
 using Humans.Application.Interfaces.Repositories;
@@ -14,7 +13,7 @@ using Humans.Application.Interfaces.Profiles;
 using Humans.Application.Interfaces.Teams;
 using Humans.Application.Interfaces.Users;
 
-namespace Humans.Application.Services.Profile;
+namespace Humans.Application.Services.Profiles;
 
 /// <summary>
 /// Service for managing account merge requests. Business logic only — no
@@ -24,7 +23,7 @@ namespace Humans.Application.Services.Profile;
 /// </summary>
 /// <remarks>
 /// Moved from <c>Humans.Infrastructure.Services</c> to
-/// <c>Humans.Application.Services.Profile</c> in PR #557 as the §15 Part 1
+/// <c>Humans.Application.Services.Profiles</c> in PR #557 as the §15 Part 1
 /// Profile-section cleanup. The owning Identity table
 /// (<c>account_merge_requests</c>) is accessed via
 /// <see cref="IAccountMergeRepository"/>; the UserEmail side uses
@@ -38,7 +37,6 @@ namespace Humans.Application.Services.Profile;
 /// <c>MergedToUserId</c> and <c>MergedAt</c> and locking out login.
 /// </para>
 /// </remarks>
-[Grandfathered("HUM0012", "Profile services live in Services.Profile but their interfaces live in Interfaces.Profiles. Plural/singular drift; folder rename deferred to a follow-up PR.", "2026-05-15", "peterdrier/Humans#557")]
 public sealed class AccountMergeService : IAccountMergeService, IUserDataContributor
 {
     private readonly IAccountMergeRepository _mergeRepository;
