+++ src/Humans.Application/Services/Profiles/EmailProblemsService.cs
@@ -1,13 +1,11 @@
-using Humans.Application.Architecture;
 using Humans.Application.DTOs.EmailProblems;
 using Humans.Application.Interfaces.Profiles;
 using Humans.Application.Interfaces.Users;
 using Humans.Domain.Helpers;
 using NodaTime;
 
-namespace Humans.Application.Services.Profile;
+namespace Humans.Application.Services.Profiles;
 
-[Grandfathered("HUM0012", "Profile services live in Services.Profile but their interfaces live in Interfaces.Profiles. Plural/singular drift; folder rename deferred to a follow-up PR.", "2026-05-15", "peterdrier/Humans#557")]
 public sealed class EmailProblemsService : IEmailProblemsService
 {
     private readonly IUserEmailService _userEmailService;
