+++ src/Humans.Infrastructure/Services/Profiles/CachingProfileService.cs
@@ -7,7 +7,7 @@ using Humans.Application.Interfaces.Caching;
 using Humans.Application.Interfaces.Repositories;
 using Humans.Domain.Entities;
 using Humans.Domain.Enums;
-using ProfilesProfileService = Humans.Application.Services.Profile.ProfileService;
+using ProfilesProfileService = Humans.Application.Services.Profiles.ProfileService;
 using Humans.Application.Interfaces.Users;
 using Humans.Application.Interfaces.Onboarding;
 using Humans.Application.Interfaces.Governance;
