+++ src/Humans.Web/Extensions/Sections/ProfileSectionExtensions.cs
@@ -5,13 +5,13 @@ using Humans.Application.Interfaces.Repositories;
 using Humans.Infrastructure.Caching;
 using Humans.Infrastructure.HostedServices;
 using Humans.Infrastructure.Services.Profiles;
-using ProfilesProfileService = Humans.Application.Services.Profile.ProfileService;
-using ProfilesContactFieldService = Humans.Application.Services.Profile.ContactFieldService;
-using ProfilesUserEmailService = Humans.Application.Services.Profile.UserEmailService;
-using ProfilesCommunicationPreferenceService = Humans.Application.Services.Profile.CommunicationPreferenceService;
-using ProfilesAccountMergeService = Humans.Application.Services.Profile.AccountMergeService;
-using ProfilesDuplicateAccountService = Humans.Application.Services.Profile.DuplicateAccountService;
-using ProfilesEmailProblemsService = Humans.Application.Services.Profile.EmailProblemsService;
+using ProfilesProfileService = Humans.Application.Services.Profiles.ProfileService;
+using ProfilesContactFieldService = Humans.Application.Services.Profiles.ContactFieldService;
+using ProfilesUserEmailService = Humans.Application.Services.Profiles.UserEmailService;
+using ProfilesCommunicationPreferenceService = Humans.Application.Services.Profiles.CommunicationPreferenceService;
+using ProfilesAccountMergeService = Humans.Application.Services.Profiles.AccountMergeService;
+using ProfilesDuplicateAccountService = Humans.Application.Services.Profiles.DuplicateAccountService;
+using ProfilesEmailProblemsService = Humans.Application.Services.Profiles.EmailProblemsService;
 using UsersAccountProvisioningService = Humans.Application.Services.Users.AccountProvisioningService;
 using UsersUserEmailProviderBackfillService = Humans.Application.Services.Users.UserEmailProviderBackfillService;
 using UsersUnsubscribeService = Humans.Application.Services.Users.UnsubscribeService;
