+++ tests/Humans.Application.Tests/Services/Gdpr/GdprExportDependencyInjectionTests.cs
@@ -11,8 +11,8 @@ using ApplicationDecisionService = Humans.Application.Services.Governance.Applic
 using BudgetService = Humans.Application.Services.Budget.BudgetService;
 using ExpenseReportService = Humans.Application.Services.Expenses.ExpenseReportService;
 using CampaignService = Humans.Application.Services.Campaigns.CampaignService;
-using ProfilesProfileService = Humans.Application.Services.Profile.ProfileService;
-using ProfilesAccountMergeService = Humans.Application.Services.Profile.AccountMergeService;
+using ProfilesProfileService = Humans.Application.Services.Profiles.ProfileService;
+using ProfilesAccountMergeService = Humans.Application.Services.Profiles.AccountMergeService;
 using UsersUserService = Humans.Application.Services.Users.UserService;
 using AuditLogService = Humans.Application.Services.AuditLog.AuditLogService;
 using CampService = Humans.Application.Services.Camps.CampService;
