+++ tests/Humans.Application.Tests/Services/Profile/CommunicationPreferenceSubscribedAtTests.cs
@@ -14,7 +14,7 @@ using NodaTime;
 using NodaTime.Testing;
 using NSubstitute;
 using Xunit;
-using CommunicationPreferenceService = Humans.Application.Services.Profile.CommunicationPreferenceService;
+using CommunicationPreferenceService = Humans.Application.Services.Profiles.CommunicationPreferenceService;
 
 namespace Humans.Application.Tests.Services.Profiles;
