+++ tests/Humans.Application.Tests/Services/UserEmailServiceTests.cs
@@ -3,7 +3,7 @@ using System;
 using Humans.Application.DTOs;
 using Humans.Application.Interfaces.AuditLog;
 using Humans.Application.Interfaces.Repositories;
-using Humans.Application.Services.Profile;
+using Humans.Application.Services.Profiles;
 using Humans.Domain.Entities;
 using Humans.Domain.Enums;
 using Humans.Testing;
